requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
pytorch-lightning>=2.0.0

# NVIDIA NeMo for FastConformer
nemo_toolkit[asr]>=1.20.0
# Alternative: use transformers if NeMo is too heavy
# transformers>=4.30.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0
resampy>=0.4.2

# Video Processing
opencv-python>=4.8.0
av>=10.0.0
decord>=0.6.0

# Vision Models
timm>=0.9.0  # For SigLip2 and other vision models
transformers>=4.30.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
wandb>=0.15.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# Utilities
tqdm>=4.65.0
einops>=0.6.1
rich>=13.0.0

# Face Detection (for ROI extraction)
mediapipe>=0.10.0
# Alternative: facenet-pytorch>=2.5.0

# Metrics
torchmetrics>=1.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
ipython>=8.14.0
jupyter>=1.0.0

# Optional: For Liquid Neural Networks
# ncps>=0.0.6  # Neural Circuit Policies

